refactor(get-message): merge condig messages nd titles
