+++ jobs/create-group-version-branch.js
@@ -5,8 +5,7 @@ const Log = require('gk-log')
 const dbs = require('../lib/dbs')
 const getConfig = require('../lib/get-config')
 const getInfos = require('../lib/get-infos')
-const getMessage = require('../lib/get-message')
-const getPRTitle = require('../lib/get-pr-title')
+const {getMessage, getPRTitle} = require('../lib/get-message')
 const createBranch = require('../lib/create-branch')
 const statsd = require('../lib/statsd')
 const env = require('../lib/env')
