+++ jobs/create-initial-branch.js
@@ -13,7 +13,7 @@ const env = require('../lib/env')
 const dbs = require('../lib/dbs')
 const getConfig = require('../lib/get-config')
 const { discoverPackageFilePaths } = require('../lib/get-files')
-const getMessage = require('../lib/get-message')
+const { getMessage } = require('../lib/get-message')
 const createBranch = require('../lib/create-branch')
 const statsd = require('../lib/statsd')
 const { updateRepoDoc } = require('../lib/repository-docs')
