+++ jobs/create-version-branch.js
@@ -4,8 +4,7 @@ const Log = require('gk-log')
 
 const dbs = require('../lib/dbs')
 const getConfig = require('../lib/get-config')
-const getMessage = require('../lib/get-message')
-const getPRTitle = require('../lib/get-pr-title')
+const {getMessage, getPRTitle} = require('../lib/get-message')
 const getInfos = require('../lib/get-infos')
 const createBranch = require('../lib/create-branch')
 const statsd = require('../lib/statsd')
