+++ lib/get-message.js
@@ -1,6 +1,6 @@
-const { defaultCommitMessages } = require('./default-config-messages')
+const { defaultCommitMessages, defaultPRTitles } = require('./default-config-messages')
 
-function replaceMessageVariables (message, variables) {
+function replaceVariables (message, variables) {
   let replacedString = message
 
   Object.keys(variables).forEach((key) => {
@@ -10,7 +10,7 @@ function replaceMessageVariables (message, variables) {
   return replacedString
 }
 
-module.exports = function (commitMessages, messageKey, values) {
+function getMessage (commitMessages, messageKey, values) {
   if (!Object.prototype.hasOwnProperty.call(commitMessages, messageKey)) {
     throw new Error(`Unknown message messageKey '${messageKey}'`)
   }
@@ -18,13 +18,35 @@ module.exports = function (commitMessages, messageKey, values) {
   // get rid of null and undefined
   const templateValues = Object.assign({}, values)
 
-  let commitMessage = replaceMessageVariables(commitMessages[messageKey], templateValues)
+  let commitMessage = replaceVariables(commitMessages[messageKey], templateValues)
 
   // if someone replaced the variable name with something else,
   // return the default message for that messageKey
   if (commitMessage.match(/\${.+?}/)) {
-    commitMessage = replaceMessageVariables(defaultCommitMessages[messageKey], templateValues)
+    commitMessage = replaceVariables(defaultCommitMessages[messageKey], templateValues)
   }
 
   return commitMessage
+}
+
+function getPRTitle ({version, dependency, group, prTitles}) {
+  const variables = {dependency, group}
+
+  if (!prTitles[version]) {
+    throw new Error('exited: Unknown PR key')
+  }
+
+  let PRTitle = replaceVariables(prTitles[version], variables)
+  // if someone replaced the variable name with something else,
+  // return the default message for that messageKey
+  if (PRTitle.match(/\${.+?}/)) {
+    PRTitle = replaceVariables(defaultPRTitles[version], variables)
+  }
+
+  return PRTitle
+}
+
+module.exports = {
+  getMessage,
+  getPRTitle
 }
