+++ lib/open-issue.js
@@ -5,7 +5,7 @@ const dbs = require('./dbs')
 const createBranch = require('./create-branch')
 const updatedAt = require('./updated-at')
 const getConfig = require('./get-config')
-const getMessage = require('./get-message')
+const { getMessage } = require('./get-message')
 const statsd = require('./statsd')
 const githubQueue = require('./github-queue')
